+++ src/vs/workbench/contrib/issue/browser/issue.ts
@@ -2,31 +2,25 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
+import { Disposable } from 'vs/base/common/lifecycle';
+import { IProductConfiguration } from 'vs/base/common/product';
 import { $, createStyleSheet, reset, windowOpenNoOpener } from 'vs/base/browser/dom';
 import { Button, unthemedButtonStyles } from 'vs/base/browser/ui/button/button';
 import { renderIcon } from 'vs/base/browser/ui/iconLabel/iconLabels';
-import { mainWindow } from 'vs/base/browser/window';
 import { Delayer, RunOnceScheduler } from 'vs/base/common/async';
 import { Codicon } from 'vs/base/common/codicons';
-import { groupBy } from 'vs/base/common/collections';
 import { debounce } from 'vs/base/common/decorators';
 import { CancellationError } from 'vs/base/common/errors';
-import { Disposable } from 'vs/base/common/lifecycle';
-import { isLinuxSnap, isMacintosh } from 'vs/base/common/platform';
+import { isLinuxSnap } from 'vs/base/common/platform';
 import { escape } from 'vs/base/common/strings';
-import { ThemeIcon } from 'vs/base/common/themables';
 import { URI } from 'vs/base/common/uri';
-import { IssueReporterModel, IssueReporterData as IssueReporterModelData } from 'vs/workbench/browser/issues/issueReporterModel';
 import { localize } from 'vs/nls';
-import { isRemoteDiagnosticError } from 'vs/platform/diagnostics/common/diagnostics';
-import { IIssueMainService, IssueReporterData, IssueReporterExtensionData, IssueReporterStyles, IssueReporterWindowConfiguration, IssueType } from 'vs/platform/issue/common/issue';
+import { IIssueMainService, IssueReporterData, IssueReporterExtensionData, IssueReporterStyles, IssueType } from 'vs/platform/issue/common/issue';
 import { normalizeGitHubUrl } from 'vs/platform/issue/common/issueReporterUtil';
-import { INativeHostService } from 'vs/platform/native/common/native';
 import { getIconsStyleSheet } from 'vs/platform/theme/browser/iconsStyleSheet';
-import { applyZoom, zoomIn, zoomOut } from 'vs/platform/window/electron-sandbox/window';
+import { IssueReporterModel, IssueReporterData as IssueReporterModelData } from 'vs/workbench/contrib/issue/browser/issueReporterModel';
+import { mainWindow } from 'vs/base/browser/window';
 
-// GitHub has let us know that we could up our limit here to 8k. We chose 7500 to play it safe.
-// ref https://github.com/microsoft/vscode/issues/159191
 const MAX_URL_LENGTH = 7500;
 
 interface SearchResult {
@@ -42,41 +36,50 @@ enum IssueSource {
 	Unknown = 'unknown'
 }
 
-export class IssueReporter extends Disposable {
-	private readonly issueReporterModel: IssueReporterModel;
-	private numberOfSearchResultsDisplayed = 0;
-	private receivedSystemInfo = false;
-	private receivedPerformanceInfo = false;
-	private shouldQueueSearch = false;
-	private hasBeenSubmitted = false;
-	private openReporter = false;
-	private loadingExtensionData = false;
-	private selectedExtension = '';
-	private delayedSubmit = new Delayer<void>(300);
-	private readonly previewButton!: Button;
-	private nonGitHubIssueUrl = false;
+
+export class BaseIssueReporterService extends Disposable {
+	public issueReporterModel: IssueReporterModel;
+	public receivedSystemInfo = false;
+	public numberOfSearchResultsDisplayed = 0;
+	public receivedPerformanceInfo = false;
+	public shouldQueueSearch = false;
+	public hasBeenSubmitted = false;
+	public openReporter = false;
+	public loadingExtensionData = false;
+	public selectedExtension = '';
+	public delayedSubmit = new Delayer<void>(300);
+	public previewButton!: Button;
+	public nonGitHubIssueUrl = false;
 
 	constructor(
-		private readonly configuration: IssueReporterWindowConfiguration,
-		@INativeHostService private readonly nativeHostService: INativeHostService,
-		@IIssueMainService private readonly issueMainService: IIssueMainService
+		public disableExtensions: boolean,
+		public data: IssueReporterData,
+		public os: {
+			type: string;
+			arch: string;
+			release: string;
+		},
+		public product: IProductConfiguration,
+		public readonly window: Window,
+		public readonly isWeb: boolean,
+		@IIssueMainService public readonly issueMainService: IIssueMainService
 	) {
 		super();
-		const targetExtension = configuration.data.extensionId ? configuration.data.enabledExtensions.find(extension => extension.id.toLocaleLowerCase() === configuration.data.extensionId?.toLocaleLowerCase()) : undefined;
+		const targetExtension = data.extensionId ? data.enabledExtensions.find(extension => extension.id.toLocaleLowerCase() === data.extensionId?.toLocaleLowerCase()) : undefined;
 		this.issueReporterModel = new IssueReporterModel({
-			...configuration.data,
-			issueType: configuration.data.issueType || IssueType.Bug,
+			...data,
+			issueType: data.issueType || IssueType.Bug,
 			versionInfo: {
-				vscodeVersion: `${configuration.product.nameShort} ${!!configuration.product.darwinUniversalAssetId ? `${configuration.product.version} (Universal)` : configuration.product.version} (${configuration.product.commit || 'Commit unknown'}, ${configuration.product.date || 'Date unknown'})`,
-				os: `${this.configuration.os.type} ${this.configuration.os.arch} ${this.configuration.os.release}${isLinuxSnap ? ' snap' : ''}`
+				vscodeVersion: `${product.nameShort} ${!!product.darwinUniversalAssetId ? `${product.version} (Universal)` : product.version} (${product.commit || 'Commit unknown'}, ${product.date || 'Date unknown'})`,
+				os: `${this.os.type} ${this.os.arch} ${this.os.release}${isLinuxSnap ? ' snap' : ''}`
 			},
-			extensionsDisabled: !!configuration.disableExtensions,
-			fileOnExtension: configuration.data.extensionId ? !targetExtension?.isBuiltin : undefined,
+			extensionsDisabled: !!this.disableExtensions,
+			fileOnExtension: data.extensionId ? !targetExtension?.isBuiltin : undefined,
 			selectedExtension: targetExtension
 		});
 
-		const fileOnMarketplace = configuration.data.issueSource === IssueSource.Marketplace;
-		const fileOnProduct = configuration.data.issueSource === IssueSource.VSCode;
+		const fileOnMarketplace = data.issueSource === IssueSource.Marketplace;
+		const fileOnProduct = data.issueSource === IssueSource.VSCode;
 		this.issueReporterModel.update({ fileOnMarketplace, fileOnProduct });
 
 		//TODO: Handle case where extension is not activated
@@ -90,7 +93,7 @@ export class IssueReporter extends Disposable {
 			this.updatePreviewButtonState();
 		}
 
-		const issueTitle = configuration.data.issueTitle;
+		const issueTitle = data.issueTitle;
 		if (issueTitle) {
 			const issueTitleElement = this.getElementById<HTMLInputElement>('issue-title');
 			if (issueTitleElement) {
@@ -98,7 +101,7 @@ export class IssueReporter extends Disposable {
 			}
 		}
 
-		const issueBody = configuration.data.issueBody;
+		const issueBody = data.issueBody;
 		if (issueBody) {
 			const description = this.getElementById<HTMLTextAreaElement>('description');
 			if (description) {
@@ -107,20 +110,7 @@ export class IssueReporter extends Disposable {
 			}
 		}
 
-		this.issueMainService.$getSystemInfo().then(info => {
-			this.issueReporterModel.update({ systemInfo: info });
-			this.receivedSystemInfo = true;
-
-			this.updateSystemInfo(this.issueReporterModel.getData());
-			this.updatePreviewButtonState();
-		});
-		if (configuration.data.issueType === IssueType.PerformanceIssue) {
-			this.issueMainService.$getPerformanceInfo().then(info => {
-				this.updatePerformanceInfo(info as Partial<IssueReporterData>);
-			});
-		}
-
-		if (mainWindow.document.documentElement.lang !== 'en') {
+		if (this.window.document.documentElement.lang !== 'en') {
 			show(this.getElementById('english'));
 		}
 
@@ -138,16 +128,10 @@ export class IssueReporter extends Disposable {
 		delayer.schedule();
 
 		this.setUpTypes();
-		this.setEventHandlers();
-		applyZoom(configuration.data.zoomLevel, mainWindow);
-		this.applyStyles(configuration.data.styles);
-		this.handleExtensionData(configuration.data.enabledExtensions);
-		this.updateExperimentsInfo(configuration.data.experiments);
-		this.updateRestrictedMode(configuration.data.restrictedMode);
-		this.updateUnsupportedMode(configuration.data.isUnsupported);
+		this.applyStyles(data.styles);
 
 		// Handle case where extension is pre-selected through the command
-		if ((configuration.data.data || configuration.data.uri) && targetExtension) {
+		if ((data.data || data.uri) && targetExtension) {
 			this.updateExtensionStatus(targetExtension);
 		}
 	}
@@ -159,10 +143,10 @@ export class IssueReporter extends Disposable {
 	setInitialFocus() {
 		const { fileOnExtension } = this.issueReporterModel.getData();
 		if (fileOnExtension) {
-			const issueTitle = mainWindow.document.getElementById('issue-title');
+			const issueTitle = this.window.document.getElementById('issue-title');
 			issueTitle?.focus();
 		} else {
-			const issueType = mainWindow.document.getElementById('issue-type');
+			const issueType = this.window.document.getElementById('issue-type');
 			issueType?.focus();
 		}
 	}
@@ -173,7 +157,7 @@ export class IssueReporter extends Disposable {
 		const content: string[] = [];
 
 		if (styles.inputBackground) {
-			content.push(`input[type="text"], textarea, select, .issues-container > .issue > .issue-state, .block-info { background-color: ${styles.inputBackground}; }`);
+			content.push(`input[type="text"], textarea, select, .issues-container > .issue > .issue-state, .block-info { background-color: ${styles.inputBackground} !important; }`);
 		}
 
 		if (styles.inputBorder) {
@@ -183,7 +167,7 @@ export class IssueReporter extends Disposable {
 		}
 
 		if (styles.inputForeground) {
-			content.push(`input[type="text"], textarea, select, .issues-container > .issue > .issue-state, .block-info { color: ${styles.inputForeground}; }`);
+			content.push(`input[type="text"], textarea, select, .issues-container > .issue > .issue-state, .block-info { color: ${styles.inputForeground} !important; }`);
 		}
 
 		if (styles.inputErrorBorder) {
@@ -240,24 +224,8 @@ export class IssueReporter extends Disposable {
 		}
 
 		styleTag.textContent = content.join('\n');
-		mainWindow.document.head.appendChild(styleTag);
-		mainWindow.document.body.style.color = styles.color || '';
-	}
-
-	private handleExtensionData(extensions: IssueReporterExtensionData[]) {
-		const installedExtensions = extensions.filter(x => !x.isBuiltin);
-		const { nonThemes, themes } = groupBy(installedExtensions, ext => {
-			return ext.isTheme ? 'themes' : 'nonThemes';
-		});
-
-		const numberOfThemeExtesions = themes && themes.length;
-		this.issueReporterModel.update({ numberOfThemeExtesions, enabledNonThemeExtesions: nonThemes, allExtensions: installedExtensions });
-		this.updateExtensionTable(nonThemes, numberOfThemeExtesions);
-		if (this.configuration.disableExtensions || installedExtensions.length === 0) {
-			(<HTMLButtonElement>this.getElementById('disableExtensions')).disabled = true;
-		}
-
-		this.updateExtensionSelector(installedExtensions);
+		this.window.document.head.appendChild(styleTag);
+		this.window.document.body.style.color = styles.color || '';
 	}
 
 	private async updateIssueReporterUri(extension: IssueReporterExtensionData): Promise<void> {
@@ -271,17 +239,7 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private async sendReporterMenu(extension: IssueReporterExtensionData): Promise<IssueReporterData | undefined> {
-		try {
-			const data = await this.issueMainService.$sendReporterMenu(extension.id, extension.name);
-			return data;
-		} catch (e) {
-			console.error(e);
-			return undefined;
-		}
-	}
-
-	private setEventHandlers(): void {
+	public setEventHandlers(): void {
 		this.addEventListener('issue-type', 'change', (event: Event) => {
 			const issueType = parseInt((<HTMLInputElement>event.target).value);
 			this.issueReporterModel.update({ issueType: issueType });
@@ -309,7 +267,7 @@ export class IssueReporter extends Disposable {
 			});
 		});
 
-		const showInfoElements = mainWindow.document.getElementsByClassName('showInfo');
+		const showInfoElements = this.window.document.getElementsByClassName('showInfo');
 		for (let i = 0; i < showInfoElements.length; i++) {
 			const showInfo = showInfoElements.item(i)!;
 			(showInfo as HTMLAnchorElement).addEventListener('click', (e: MouseEvent) => {
@@ -395,78 +353,9 @@ export class IssueReporter extends Disposable {
 			const { fileOnExtension, fileOnMarketplace } = this.issueReporterModel.getData();
 			this.searchIssues(title, fileOnExtension, fileOnMarketplace);
 		});
-
-		this.previewButton.onDidClick(async () => {
-			this.delayedSubmit.trigger(async () => {
-				this.createIssue();
-			});
-		});
-
-		this.addEventListener('disableExtensions', 'click', () => {
-			this.issueMainService.$reloadWithExtensionsDisabled();
-		});
-
-		this.addEventListener('extensionBugsLink', 'click', (e: Event) => {
-			const url = (<HTMLElement>e.target).innerText;
-			windowOpenNoOpener(url);
-		});
-
-		this.addEventListener('disableExtensions', 'keydown', (e: Event) => {
-			e.stopPropagation();
-			if ((e as KeyboardEvent).keyCode === 13 || (e as KeyboardEvent).keyCode === 32) {
-				this.issueMainService.$reloadWithExtensionsDisabled();
-			}
-		});
-
-		mainWindow.document.onkeydown = async (e: KeyboardEvent) => {
-			const cmdOrCtrlKey = isMacintosh ? e.metaKey : e.ctrlKey;
-			// Cmd/Ctrl+Enter previews issue and closes window
-			if (cmdOrCtrlKey && e.keyCode === 13) {
-				this.delayedSubmit.trigger(async () => {
-					if (await this.createIssue()) {
-						this.close();
-					}
-				});
-			}
-
-			// Cmd/Ctrl + w closes issue window
-			if (cmdOrCtrlKey && e.keyCode === 87) {
-				e.stopPropagation();
-				e.preventDefault();
-
-				const issueTitle = (<HTMLInputElement>this.getElementById('issue-title'))!.value;
-				const { issueDescription } = this.issueReporterModel.getData();
-				if (!this.hasBeenSubmitted && (issueTitle || issueDescription)) {
-					// fire and forget
-					this.issueMainService.$showConfirmCloseDialog();
-				} else {
-					this.close();
-				}
-			}
-
-			// Cmd/Ctrl + zooms in
-			if (cmdOrCtrlKey && e.keyCode === 187) {
-				zoomIn(mainWindow);
-			}
-
-			// Cmd/Ctrl - zooms out
-			if (cmdOrCtrlKey && e.keyCode === 189) {
-				zoomOut(mainWindow);
-			}
-
-			// With latest electron upgrade, cmd+a is no longer propagating correctly for inputs in this window on mac
-			// Manually perform the selection
-			if (isMacintosh) {
-				if (cmdOrCtrlKey && e.keyCode === 65 && e.target) {
-					if (e.target instanceof HTMLInputElement || e.target instanceof HTMLTextAreaElement) {
-						(<HTMLInputElement>e.target).select();
-					}
-				}
-			}
-		};
 	}
 
-	private updatePerformanceInfo(info: Partial<IssueReporterData>) {
+	public updatePerformanceInfo(info: Partial<IssueReporterData>) {
 		this.issueReporterModel.update(info);
 		this.receivedPerformanceInfo = true;
 
@@ -476,9 +365,9 @@ export class IssueReporter extends Disposable {
 		this.updatePreviewButtonState();
 	}
 
-	private updatePreviewButtonState() {
+	public updatePreviewButtonState() {
 		if (this.isPreviewEnabled()) {
-			if (this.configuration.data.githubAccessToken) {
+			if (this.data.githubAccessToken) {
 				this.previewButton.label = localize('createOnGitHub', "Create on GitHub");
 			} else {
 				this.previewButton.label = localize('previewOnGitHub', "Preview on GitHub");
@@ -525,16 +414,18 @@ export class IssueReporter extends Disposable {
 			return false;
 		}
 
-		if (issueType === IssueType.Bug && this.receivedSystemInfo) {
-			return true;
-		}
+		if (this.isWeb) {
+			if (issueType === IssueType.FeatureRequest || issueType === IssueType.PerformanceIssue || issueType === IssueType.Bug) {
+				return true;
+			}
+		} else {
+			if (issueType === IssueType.Bug && this.receivedSystemInfo) {
+				return true;
+			}
 
-		if (issueType === IssueType.PerformanceIssue && this.receivedSystemInfo && this.receivedPerformanceInfo) {
-			return true;
-		}
-
-		if (issueType === IssueType.FeatureRequest) {
-			return true;
+			if (issueType === IssueType.PerformanceIssue && this.receivedSystemInfo && this.receivedPerformanceInfo) {
+				return true;
+			}
 		}
 
 		return false;
@@ -545,12 +436,12 @@ export class IssueReporter extends Disposable {
 		return selectedExtension && selectedExtension.repositoryUrl;
 	}
 
-	private getExtensionBugsUrl(): string | undefined {
+	public getExtensionBugsUrl(): string | undefined {
 		const selectedExtension = this.issueReporterModel.getData().selectedExtension;
 		return selectedExtension && selectedExtension.bugsUrl;
 	}
 
-	private searchVSCodeIssues(title: string, issueDescription?: string): void {
+	public searchVSCodeIssues(title: string, issueDescription?: string): void {
 		if (title) {
 			this.searchDuplicates(title, issueDescription);
 		} else {
@@ -558,7 +449,7 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private searchIssues(title: string, fileOnExtension: boolean | undefined, fileOnMarketplace: boolean | undefined): void {
+	public searchIssues(title: string, fileOnExtension: boolean | undefined, fileOnMarketplace: boolean | undefined): void {
 		if (fileOnExtension) {
 			return this.searchExtensionIssues(title);
 		}
@@ -593,18 +484,18 @@ export class IssueReporter extends Disposable {
 
 	private searchMarketplaceIssues(title: string): void {
 		if (title) {
-			const gitHubInfo = this.parseGitHubUrl(this.configuration.product.reportMarketplaceIssueUrl!);
+			const gitHubInfo = this.parseGitHubUrl(this.product.reportMarketplaceIssueUrl!);
 			if (gitHubInfo) {
 				return this.searchGitHub(`${gitHubInfo.owner}/${gitHubInfo.repositoryName}`, title);
 			}
 		}
 	}
 
-	private async close(): Promise<void> {
+	public async close(): Promise<void> {
 		await this.issueMainService.$closeReporter();
 	}
 
-	private clearSearchResults(): void {
+	public clearSearchResults(): void {
 		const similarIssues = this.getElementById('similar-issues')!;
 		similarIssues.innerText = '';
 		this.numberOfSearchResultsDisplayed = 0;
@@ -637,10 +528,10 @@ export class IssueReporter extends Disposable {
 					}
 				}
 			}).catch(_ => {
-				// Ignore
+				console.warn('Timeout or query limit exceeded');
 			});
 		}).catch(_ => {
-			// Ignore
+			console.warn('Error fetching GitHub issues');
 		});
 	}
 
@@ -739,7 +630,7 @@ export class IssueReporter extends Disposable {
 		this.setSourceOptions();
 	}
 
-	private makeOption(value: string, description: string, disabled: boolean): HTMLOptionElement {
+	public makeOption(value: string, description: string, disabled: boolean): HTMLOptionElement {
 		const option: HTMLOptionElement = document.createElement('option');
 		option.disabled = disabled;
 		option.value = value;
@@ -748,7 +639,7 @@ export class IssueReporter extends Disposable {
 		return option;
 	}
 
-	private setSourceOptions(): void {
+	public setSourceOptions(): void {
 		const sourceSelect = this.getElementById('issue-source')! as HTMLSelectElement;
 		const { issueType, fileOnExtension, selectedExtension, fileOnMarketplace, fileOnProduct } = this.issueReporterModel.getData();
 		let selected = sourceSelect.selectedIndex;
@@ -768,7 +659,7 @@ export class IssueReporter extends Disposable {
 		sourceSelect.append(this.makeOption('', localize('selectSource', "Select source"), true));
 		sourceSelect.append(this.makeOption(IssueSource.VSCode, localize('vscode', "Visual Studio Code"), false));
 		sourceSelect.append(this.makeOption(IssueSource.Extension, localize('extension', "A VS Code extension"), false));
-		if (this.configuration.product.reportMarketplaceIssueUrl) {
+		if (this.product.reportMarketplaceIssueUrl) {
 			sourceSelect.append(this.makeOption(IssueSource.Marketplace, localize('marketplace', "Extensions Marketplace"), false));
 		}
 
@@ -784,16 +675,16 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private renderBlocks(): void {
+	public renderBlocks(): void {
 		// Depending on Issue Type, we render different blocks and text
 		const { issueType, fileOnExtension, fileOnMarketplace, selectedExtension } = this.issueReporterModel.getData();
 		const blockContainer = this.getElementById('block-container');
-		const systemBlock = mainWindow.document.querySelector('.block-system');
-		const processBlock = mainWindow.document.querySelector('.block-process');
-		const workspaceBlock = mainWindow.document.querySelector('.block-workspace');
-		const extensionsBlock = mainWindow.document.querySelector('.block-extensions');
-		const experimentsBlock = mainWindow.document.querySelector('.block-experiments');
-		const extensionDataBlock = mainWindow.document.querySelector('.block-extension-data');
+		const systemBlock = this.window.document.querySelector('.block-system');
+		const processBlock = this.window.document.querySelector('.block-process');
+		const workspaceBlock = this.window.document.querySelector('.block-workspace');
+		const extensionsBlock = this.window.document.querySelector('.block-extensions');
+		const experimentsBlock = this.window.document.querySelector('.block-experiments');
+		const extensionDataBlock = this.window.document.querySelector('.block-extension-data');
 
 		const problemSource = this.getElementById('problem-source')!;
 		const descriptionTitle = this.getElementById('issue-description-label')!;
@@ -849,6 +740,7 @@ export class IssueReporter extends Disposable {
 					show(extensionDataBlock);
 				}
 			}, 100);
+			show(extensionDataBlock);
 		}
 
 		if (issueType === IssueType.Bug) {
@@ -886,7 +778,7 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private validateInput(inputId: string): boolean {
+	public validateInput(inputId: string): boolean {
 		const inputElement = (<HTMLInputElement>this.getElementById(inputId));
 		const inputValidationMessage = this.getElementById(`${inputId}-empty-error`);
 		const descriptionShortMessage = this.getElementById(`description-short-error`);
@@ -911,7 +803,7 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private validateInputs(): boolean {
+	public validateInputs(): boolean {
 		let isValid = true;
 		['issue-title', 'description', 'issue-source'].forEach(elementId => {
 			isValid = this.validateInput(elementId) && isValid;
@@ -924,7 +816,7 @@ export class IssueReporter extends Disposable {
 		return isValid;
 	}
 
-	private async submitToGitHub(issueTitle: string, issueBody: string, gitHubDetails: { owner: string; repositoryName: string }): Promise<boolean> {
+	public async submitToGitHub(issueTitle: string, issueBody: string, gitHubDetails: { owner: string; repositoryName: string }): Promise<boolean> {
 		const url = `https://api.github.com/repos/${gitHubDetails.owner}/${gitHubDetails.repositoryName}/issues`;
 		const init = {
 			method: 'POST',
@@ -934,7 +826,8 @@ export class IssueReporter extends Disposable {
 			}),
 			headers: new Headers({
 				'Content-Type': 'application/json',
-				'Authorization': `Bearer ${this.configuration.data.githubAccessToken}`
+				'Authorization': `Bearer ${this.data.githubAccessToken}`,
+				'User-Agent': 'request'
 			})
 		};
 
@@ -944,12 +837,12 @@ export class IssueReporter extends Disposable {
 			return false;
 		}
 		const result = await response.json();
-		await this.nativeHostService.openExternal(result.html_url);
+		mainWindow.open(result.html_url, '_blank');
 		this.close();
 		return true;
 	}
 
-	private async createIssue(): Promise<boolean> {
+	public async createIssue(): Promise<boolean> {
 		const selectedExtension = this.issueReporterModel.getData().selectedExtension;
 		const hasUri = this.nonGitHubIssueUrl;
 		// Short circuit if the extension provides a custom issue handler
@@ -957,7 +850,6 @@ export class IssueReporter extends Disposable {
 			const url = this.getExtensionBugsUrl();
 			if (url) {
 				this.hasBeenSubmitted = true;
-				await this.nativeHostService.openExternal(url);
 				return true;
 			}
 		}
@@ -965,7 +857,7 @@ export class IssueReporter extends Disposable {
 		if (!this.validateInputs()) {
 			// If inputs are invalid, set focus to the first one and add listeners on them
 			// to detect further changes
-			const invalidInput = mainWindow.document.getElementsByClassName('invalid-input');
+			const invalidInput = this.window.document.getElementsByClassName('invalid-input');
 			if (invalidInput.length) {
 				(<HTMLInputElement>invalidInput[0]).focus();
 			}
@@ -1008,7 +900,7 @@ export class IssueReporter extends Disposable {
 		}
 
 		const gitHubDetails = this.parseGitHubUrl(issueUrl);
-		if (this.configuration.data.githubAccessToken && gitHubDetails) {
+		if (this.data.githubAccessToken && gitHubDetails) {
 			return this.submitToGitHub(issueTitle, issueBody, gitHubDetails);
 		}
 
@@ -1024,30 +916,29 @@ export class IssueReporter extends Disposable {
 			}
 		}
 
-		await this.nativeHostService.openExternal(url);
+		this.window.open(url, '_blank');
+
 		return true;
 	}
 
-	private async writeToClipboard(baseUrl: string, issueBody: string): Promise<string> {
+	public async writeToClipboard(baseUrl: string, issueBody: string): Promise<string> {
 		const shouldWrite = await this.issueMainService.$showClipboardDialog();
 		if (!shouldWrite) {
 			throw new CancellationError();
 		}
 
-		await this.nativeHostService.writeClipboardText(issueBody);
-
 		return baseUrl + `&body=${encodeURIComponent(localize('pasteData', "We have written the needed data into your clipboard because it was too large to send. Please paste."))}`;
 	}
 
-	private getIssueUrl(): string {
+	public getIssueUrl(): string {
 		return this.issueReporterModel.fileOnExtension()
 			? this.getExtensionGitHubUrl()
 			: this.issueReporterModel.getData().fileOnMarketplace
-				? this.configuration.product.reportMarketplaceIssueUrl!
-				: this.configuration.product.reportIssueUrl!;
+				? this.product.reportMarketplaceIssueUrl!
+				: this.product.reportIssueUrl!;
 	}
 
-	private parseGitHubUrl(url: string): undefined | { repositoryName: string; owner: string } {
+	public parseGitHubUrl(url: string): undefined | { repositoryName: string; owner: string } {
 		// Assumes a GitHub url to a particular repo, https://github.com/repositoryName/owner.
 		// Repository name and owner cannot contain '/'
 		const match = /^https?:\/\/github\.com\/([^\/]*)\/([^\/]*).*/.exec(url);
@@ -1082,7 +973,7 @@ export class IssueReporter extends Disposable {
 		return repositoryUrl;
 	}
 
-	private getIssueUrlWithTitle(issueTitle: string, repositoryUrl: string): string {
+	public getIssueUrlWithTitle(issueTitle: string, repositoryUrl: string): string {
 		if (this.issueReporterModel.fileOnExtension()) {
 			repositoryUrl = repositoryUrl + '/issues/new';
 		}
@@ -1091,198 +982,19 @@ export class IssueReporter extends Disposable {
 		return `${repositoryUrl}${queryStringPrefix}title=${encodeURIComponent(issueTitle)}`;
 	}
 
-	private updateSystemInfo(state: IssueReporterModelData) {
-		const target = mainWindow.document.querySelector<HTMLElement>('.block-system .block-info');
-
-		if (target) {
-			const systemInfo = state.systemInfo!;
-			const renderedDataTable = $('table', undefined,
-				$('tr', undefined,
-					$('td', undefined, 'CPUs'),
-					$('td', undefined, systemInfo.cpus || '')
-				),
-				$('tr', undefined,
-					$('td', undefined, 'GPU Status' as string),
-					$('td', undefined, Object.keys(systemInfo.gpuStatus).map(key => `${key}: ${systemInfo.gpuStatus[key]}`).join('\n'))
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Load (avg)' as string),
-					$('td', undefined, systemInfo.load || '')
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Memory (System)' as string),
-					$('td', undefined, systemInfo.memory)
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Process Argv' as string),
-					$('td', undefined, systemInfo.processArgs)
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Screen Reader' as string),
-					$('td', undefined, systemInfo.screenReader)
-				),
-				$('tr', undefined,
-					$('td', undefined, 'VM'),
-					$('td', undefined, systemInfo.vmHint)
-				)
-			);
-			reset(target, renderedDataTable);
-
-			systemInfo.remoteData.forEach(remote => {
-				target.appendChild($<HTMLHRElement>('hr'));
-				if (isRemoteDiagnosticError(remote)) {
-					const remoteDataTable = $('table', undefined,
-						$('tr', undefined,
-							$('td', undefined, 'Remote'),
-							$('td', undefined, remote.hostName)
-						),
-						$('tr', undefined,
-							$('td', undefined, ''),
-							$('td', undefined, remote.errorMessage)
-						)
-					);
-					target.appendChild(remoteDataTable);
-				} else {
-					const remoteDataTable = $('table', undefined,
-						$('tr', undefined,
-							$('td', undefined, 'Remote'),
-							$('td', undefined, remote.latency ? `${remote.hostName} (latency: ${remote.latency.current.toFixed(2)}ms last, ${remote.latency.average.toFixed(2)}ms average)` : remote.hostName)
-						),
-						$('tr', undefined,
-							$('td', undefined, 'OS'),
-							$('td', undefined, remote.machineInfo.os)
-						),
-						$('tr', undefined,
-							$('td', undefined, 'CPUs'),
-							$('td', undefined, remote.machineInfo.cpus || '')
-						),
-						$('tr', undefined,
-							$('td', undefined, 'Memory (System)' as string),
-							$('td', undefined, remote.machineInfo.memory)
-						),
-						$('tr', undefined,
-							$('td', undefined, 'VM'),
-							$('td', undefined, remote.machineInfo.vmHint)
-						)
-					);
-					target.appendChild(remoteDataTable);
-				}
-			});
-		}
-	}
-
-	private updateExtensionSelector(extensions: IssueReporterExtensionData[]): void {
-		interface IOption {
-			name: string;
-			id: string;
-		}
-
-		const extensionOptions: IOption[] = extensions.map(extension => {
-			return {
-				name: extension.displayName || extension.name || '',
-				id: extension.id
-			};
-		});
-
-		// Sort extensions by name
-		extensionOptions.sort((a, b) => {
-			const aName = a.name.toLowerCase();
-			const bName = b.name.toLowerCase();
-			if (aName > bName) {
-				return 1;
-			}
-
-			if (aName < bName) {
-				return -1;
-			}
-
-			return 0;
-		});
-
-		const makeOption = (extension: IOption, selectedExtension?: IssueReporterExtensionData): HTMLOptionElement => {
-			const selected = selectedExtension && extension.id === selectedExtension.id;
-			return $<HTMLOptionElement>('option', {
-				'value': extension.id,
-				'selected': selected || ''
-			}, extension.name);
-		};
-
-		const extensionsSelector = this.getElementById<HTMLSelectElement>('extension-selector');
-		if (extensionsSelector) {
-			const { selectedExtension } = this.issueReporterModel.getData();
-			reset(extensionsSelector, this.makeOption('', localize('selectExtension', "Select extension"), true), ...extensionOptions.map(extension => makeOption(extension, selectedExtension)));
-
-			if (!selectedExtension) {
-				extensionsSelector.selectedIndex = 0;
-			}
-
-			this.addEventListener('extension-selector', 'change', async (e: Event) => {
-				this.clearExtensionData();
-				const selectedExtensionId = (<HTMLInputElement>e.target).value;
-				this.selectedExtension = selectedExtensionId;
-				const extensions = this.issueReporterModel.getData().allExtensions;
-				const matches = extensions.filter(extension => extension.id === selectedExtensionId);
-				if (matches.length) {
-					this.issueReporterModel.update({ selectedExtension: matches[0] });
-					const selectedExtension = this.issueReporterModel.getData().selectedExtension;
-					if (selectedExtension) {
-						const iconElement = document.createElement('span');
-						iconElement.classList.add(...ThemeIcon.asClassNameArray(Codicon.loading), 'codicon-modifier-spin');
-						this.setLoading(iconElement);
-						const openReporterData = await this.sendReporterMenu(selectedExtension);
-						if (openReporterData) {
-							if (this.selectedExtension === selectedExtensionId) {
-								this.removeLoading(iconElement, true);
-								this.configuration.data = openReporterData;
-							} else if (this.selectedExtension !== selectedExtensionId) {
-							}
-						}
-						else {
-							if (!this.loadingExtensionData) {
-								iconElement.classList.remove(...ThemeIcon.asClassNameArray(Codicon.loading), 'codicon-modifier-spin');
-							}
-							this.removeLoading(iconElement);
-							// if not using command, should have no configuration data in fields we care about and check later.
-							this.clearExtensionData();
-
-							// case when previous extension was opened from normal openIssueReporter command
-							selectedExtension.data = undefined;
-							selectedExtension.uri = undefined;
-						}
-						if (this.selectedExtension === selectedExtensionId) {
-							// repopulates the fields with the new data given the selected extension.
-							this.updateExtensionStatus(matches[0]);
-							this.openReporter = false;
-						}
-					} else {
-						this.issueReporterModel.update({ selectedExtension: undefined });
-						this.clearSearchResults();
-						this.clearExtensionData();
-						this.validateSelectedExtension();
-						this.updateExtensionStatus(matches[0]);
-					}
-				}
-			});
-		}
-
-		this.addEventListener('problem-source', 'change', (_) => {
-			this.validateSelectedExtension();
-		});
-	}
-
-	private clearExtensionData(): void {
+	public clearExtensionData(): void {
 		this.nonGitHubIssueUrl = false;
 		this.issueReporterModel.update({ extensionData: undefined });
-		this.configuration.data.issueBody = undefined;
-		this.configuration.data.data = undefined;
-		this.configuration.data.uri = undefined;
+		this.data.issueBody = undefined;
+		this.data.data = undefined;
+		this.data.uri = undefined;
 	}
 
-	private async updateExtensionStatus(extension: IssueReporterExtensionData) {
+	public async updateExtensionStatus(extension: IssueReporterExtensionData) {
 		this.issueReporterModel.update({ selectedExtension: extension });
 
 		// uses this.configuuration.data to ensure that data is coming from `openReporter` command.
-		const template = this.configuration.data.issueBody;
+		const template = this.data.issueBody;
 		if (template) {
 			const descriptionTextArea = this.getElementById('description')!;
 			const descriptionText = (descriptionTextArea as HTMLTextAreaElement).value;
@@ -1293,16 +1005,16 @@ export class IssueReporter extends Disposable {
 			}
 		}
 
-		const data = this.configuration.data.data;
+		const data = this.data.data;
 		if (data) {
 			this.issueReporterModel.update({ extensionData: data });
 			extension.data = data;
-			const extensionDataBlock = mainWindow.document.querySelector('.block-extension-data')!;
+			const extensionDataBlock = this.window.document.querySelector('.block-extension-data')!;
 			show(extensionDataBlock);
 			this.renderBlocks();
 		}
 
-		const uri = this.configuration.data.uri;
+		const uri = this.data.uri;
 		if (uri) {
 			extension.uri = uri;
 			this.updateIssueReporterUri(extension);
@@ -1316,7 +1028,7 @@ export class IssueReporter extends Disposable {
 		this.renderBlocks();
 	}
 
-	private validateSelectedExtension(): void {
+	public validateSelectedExtension(): void {
 		const extensionValidationMessage = this.getElementById('extension-selection-validation-error')!;
 		const extensionValidationNoUrlsMessage = this.getElementById('extension-selection-validation-error-no-url')!;
 		hide(extensionValidationMessage);
@@ -1341,7 +1053,7 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private setLoading(element: HTMLElement) {
+	public setLoading(element: HTMLElement) {
 		// Show loading
 		this.openReporter = true;
 		this.loadingExtensionData = true;
@@ -1350,7 +1062,7 @@ export class IssueReporter extends Disposable {
 		const extensionDataCaption = this.getElementById('extension-id')!;
 		hide(extensionDataCaption);
 
-		const extensionDataCaption2 = Array.from(mainWindow.document.querySelectorAll('.ext-parens'));
+		const extensionDataCaption2 = Array.from(this.window.document.querySelectorAll('.ext-parens'));
 		extensionDataCaption2.forEach(extensionDataCaption2 => hide(extensionDataCaption2));
 
 		const showLoading = this.getElementById('ext-loading')!;
@@ -1363,7 +1075,7 @@ export class IssueReporter extends Disposable {
 		this.renderBlocks();
 	}
 
-	private removeLoading(element: HTMLElement, fromReporter: boolean = false) {
+	public removeLoading(element: HTMLElement, fromReporter: boolean = false) {
 		this.openReporter = fromReporter;
 		this.loadingExtensionData = false;
 		this.updatePreviewButtonState();
@@ -1371,7 +1083,7 @@ export class IssueReporter extends Disposable {
 		const extensionDataCaption = this.getElementById('extension-id')!;
 		show(extensionDataCaption);
 
-		const extensionDataCaption2 = Array.from(mainWindow.document.querySelectorAll('.ext-parens'));
+		const extensionDataCaption2 = Array.from(this.window.document.querySelectorAll('.ext-parens'));
 		extensionDataCaption2.forEach(extensionDataCaption2 => show(extensionDataCaption2));
 
 		const hideLoading = this.getElementById('ext-loading')!;
@@ -1405,20 +1117,20 @@ export class IssueReporter extends Disposable {
 	}
 
 	private updateProcessInfo(state: IssueReporterModelData) {
-		const target = mainWindow.document.querySelector('.block-process .block-info') as HTMLElement;
+		const target = this.window.document.querySelector('.block-process .block-info') as HTMLElement;
 		if (target) {
 			reset(target, $('code', undefined, state.processInfo ?? ''));
 		}
 	}
 
 	private updateWorkspaceInfo(state: IssueReporterModelData) {
-		mainWindow.document.querySelector('.block-workspace .block-info code')!.textContent = '\n' + state.workspaceInfo;
+		this.window.document.querySelector('.block-workspace .block-info code')!.textContent = '\n' + state.workspaceInfo;
 	}
 
-	private updateExtensionTable(extensions: IssueReporterExtensionData[], numThemeExtensions: number): void {
-		const target = mainWindow.document.querySelector<HTMLElement>('.block-extensions .block-info');
+	public updateExtensionTable(extensions: IssueReporterExtensionData[], numThemeExtensions: number): void {
+		const target = this.window.document.querySelector<HTMLElement>('.block-extensions .block-info');
 		if (target) {
-			if (this.configuration.disableExtensions) {
+			if (this.disableExtensions) {
 				reset(target, localize('disabledExtensions', "Extensions are disabled"));
 				return;
 			}
@@ -1432,22 +1144,6 @@ export class IssueReporter extends Disposable {
 			}
 
 			reset(target, this.getExtensionTableHtml(extensions), document.createTextNode(themeExclusionStr));
-		}
-	}
-
-	private updateRestrictedMode(restrictedMode: boolean) {
-		this.issueReporterModel.update({ restrictedMode });
-	}
-
-	private updateUnsupportedMode(isUnsupported: boolean) {
-		this.issueReporterModel.update({ isUnsupported });
-	}
-
-	private updateExperimentsInfo(experimentInfo: string | undefined) {
-		this.issueReporterModel.update({ experimentInfo });
-		const target = mainWindow.document.querySelector<HTMLElement>('.block-experiments .block-info');
-		if (target) {
-			target.textContent = experimentInfo ? experimentInfo : localize('noCurrentExperiments', "No current experiments.");
 		}
 	}
 
@@ -1475,8 +1171,8 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private getElementById<T extends HTMLElement = HTMLElement>(elementId: string): T | undefined {
-		const element = mainWindow.document.getElementById(elementId) as T | undefined;
+	public getElementById<T extends HTMLElement = HTMLElement>(elementId: string): T | undefined {
+		const element = this.window.document.getElementById(elementId) as T | undefined;
 		if (element) {
 			return element;
 		} else {
@@ -1484,7 +1180,7 @@ export class IssueReporter extends Disposable {
 		}
 	}
 
-	private addEventListener(elementId: string, eventType: string, handler: (event: Event) => void): void {
+	public addEventListener(elementId: string, eventType: string, handler: (event: Event) => void): void {
 		const element = this.getElementById(elementId);
 		element?.addEventListener(eventType, handler);
 	}
@@ -1498,3 +1194,5 @@ export function hide(el: Element | undefined | null) {
 export function show(el: Element | undefined | null) {
 	el?.classList.remove('hidden');
 }
+
+
