+++ src/vs/workbench/contrib/issue/browser/issueReporterService.ts
@@ -10,7 +10,7 @@ import { IProductConfiguration } from 'vs/base/common/product';
 import { ThemeIcon } from 'vs/base/common/themables';
 import { localize } from 'vs/nls';
 import { IIssueMainService, IssueReporterData, IssueReporterExtensionData } from 'vs/platform/issue/common/issue';
-import { BaseIssueReporterService } from 'vs/workbench/browser/issues/issue';
+import { BaseIssueReporterService } from 'vs/workbench/contrib/issue/browser/issue';
 
 // GitHub has let us know that we could up our limit here to 8k. We chose 7500 to play it safe.
 // ref https://github.com/microsoft/vscode/issues/159191
