+++ src/vs/workbench/contrib/issue/electron-sandbox/issueReporter-dev.html
