+++ src/vs/workbench/contrib/issue/electron-sandbox/media/issueReporter.css
